pyproject.toml
[project]
name = "claude-agent-sdk-test"
version = "0.1.0"
description = "Claude Agent SDK 学习和测试项目"
readme = "README.md"
requires-python = ">=3.12"
authors = [
    { name = "ChildhoodAndy", email = "dabing1022@gmail.com" }
]
dependencies = [
    "claude-agent-sdk>=0.1.17",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[build-system]
requires = ["setuptools>=68.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.black]
line-length = 100
target-version = ["py312"]

[tool.ruff]
line-length = 100
target-version = "py312"
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
